ecpg: Don't support nested comments as per C.
Not sure why this was supported, but it broke in cases like:

   /************************/
   /* Some comment
   /************************/
   struct T {
      ...
   };
   /* next comment */

Where the non-comment struct T definition would be emitted in-place
without further parsing as if it were a comment.   This is in contrast
to C syntax where */ ends a block comment independent of whether there
are additional /* within it.