ecpg: Support [[ecpg::default]] field attribute to auto-default SERIAL fields and new EXPLICIT keyword to override.
Fields within known structs may be marked [[ecpg::default]] to indicate
that PostgreSQL's 'default' token should be passed rather than the
actual value stored in the field when performing an INSERT or UPDATE.
This provides convenient, natural, handling for struct-based operations
where the struct maps to a table's tuple type.

The behavior can be vetoed if 'EXPLICIT' or 'EXPLICIT IF (C-expr)' are
specified after the VALUES keyword.  In that case, the original
behavior, as if [[ecpg::default]] was not specified for the field, will
apply.  I.e. the fields value will be written for that column.

```c
   $struct Message {
      int id [[ecpg::default]];
      string* text;
   };

   void test()
   {
      $create temp table struct_test (id SERIAL PRIMARY KEY, text TEXT);

      $begin declare section;
      struct Message msg = {0, "some text"};
      $end declare section;

      $insert into struct_test values (:msg) returning id into :msg.id;
      /* msg.id will be 1 due to initial SERIAL value */

      msg.id = 7;
      msg.text = "some more text";
      $insert into struct_test values explicit (:msg);

      msg.id = 0;
      msg.text = "a third string";
      $insert into struct_test values explicit if (msg.id > 0) (:msg);
      /* id column of inserted record will be 2 */
   }
```